finalização do crud completo, com refatorações nos métodos atualizar e delete.

+++ src/main/java/com/example/projetolex/service/EscritorService.java
@@ -1,13 +1,16 @@
 package com.example.projetolex.service;
 
 import com.example.projetolex.domain.Escritor;
+import com.example.projetolex.domain.EscritorDTO;
 import com.example.projetolex.execption.DadoInvalidoException;
 import com.example.projetolex.repository.EscritorRepository;
+import com.example.projetolex.util.ValidadorDados;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+import static com.example.projetolex.util.ValidadorDados.*;
 import static com.example.projetolex.util.ValidadorDados.validarCpf;
 import static com.example.projetolex.util.ValidadorDados.validarEmail;
 import static com.example.projetolex.util.ValidadorDados.validarNome;
@@ -25,8 +28,24 @@ public class EscritorService {
         return escritor;
     }
 
-    public void removerEscritor(int id) {
+    public Escritor editarEscritor(EscritorDTO dto) {
+        Escritor escritor = buscarEscritorPorId(dto.getId());
+        if (escritor == null) throw new DadoInvalidoException("Escritor não encontrado.");
+        if (dto.getCpf() != null) throw new DadoInvalidoException("CPF não pode ser alterado.");
+        if (dto.getNome() != null && !validarNome(dto.getNome()) && dto.getNome() != null) throw new DadoInvalidoException("Nome inválido.");
+        if (dto.getEmail() != null && !validarEmail(dto.getEmail())) throw new DadoInvalidoException("E-mail inválido.");
+        if (dto.getIdade() != null && !validarNumero(dto.getIdade())) throw new DadoInvalidoException("Idade inválida.");
+
+        escritor.setNome(dto.getNome());
+        escritor.setIdade(dto.getIdade());
+        salvarEscritor(escritor);
+        return escritor;
+    }
+
+    public String removerEscritor(int id) {
+        if (buscarEscritorPorId(id) == null) throw new DadoInvalidoException("Escritor não encontrado.");
         repository.deleteById(id);
+        return "Escritor removido com sucesso!";
     }
 
     public Escritor buscarEscritorPorId(int id) {
